Move ImageResponse to next/og
