+++ packages/next/og.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/server/og/image-response')
