+++ test/e2e/app-dir/metadata-dynamic-routes/app/(group)/dynamic/[size]/apple-icon.tsx
@@ -1,4 +1,4 @@
-import { ImageResponse } from 'next/server'
+import { ImageResponse } from 'next/og'
 
 // without id
 export async function generateImageMetadata({ params }) {
