+++ test/e2e/app-dir/metadata-dynamic-routes/app/font/opengraph-image.tsx
@@ -1,4 +1,4 @@
-import { ImageResponse } from 'next/server'
+import { ImageResponse } from 'next/og'
 
 export const contentType = 'image/png'
