+++ test/e2e/app-dir/metadata-dynamic-routes/app/font/opengraph-image2.tsx
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 import { fileURLToPath } from 'url'
-import { ImageResponse } from 'next/server'
+import { ImageResponse } from 'next/og'
 
 export const contentType = 'image/png'
